refactor: reorganize date picker support code
