+++ fluent/src/commonMain/kotlin/io/github/composefluent/platform/PlatformCalendar.kt
@@ -1,4 +1,4 @@
-package io.github.composefluent.component
+package io.github.composefluent.platform
 
 internal enum class LocalMonthNameStyle {
     Short,
