+++ fluent/src/jsMain/kotlin/io/github/composefluent/platform/PlatformCalendar.js.kt
@@ -1,4 +1,4 @@
-package io.github.composefluent.component
+package io.github.composefluent.platform
 
 internal actual fun getLocalLocaleIdentifier(): String =
     js("navigator.language") as String
