+++ fluent/src/wasmJsMain/kotlin/io/github/composefluent/platform/PlatformCalendar.wasmJs.kt
@@ -1,6 +1,6 @@
 @file:OptIn(kotlin.js.ExperimentalWasmJsInterop::class)
 
-package io.github.composefluent.component
+package io.github.composefluent.platform
 
 @JsFun("() => navigator.language")
 private external fun getLocalLocaleIdentifierJs(): String
