+++ fluent/src/webMain/kotlin/io/github/composefluent/platform/PlatformCalendar.web.kt
@@ -1,4 +1,4 @@
-package io.github.composefluent.component
+package io.github.composefluent.platform
 
 private val localeToFirstDayMap by lazy {
     mapOf(
